Modify start screen so Space can start game too

+++ PulseTest/Assets/Scripts/MC Scripts/MentalState.cs
@@ -101,7 +101,7 @@ public class MentalState : MonoBehaviour
             }
         }
             
-        return average;
+        return (int)average;
 
 
         return 0;
